Moving problems to proper package
