+++ src/mx/grekz/leetcode/medium/E002_AddTwoNumbers.java
@@ -1,4 +1,4 @@
-package mx.grekz.leetcode.easy;
+package mx.grekz.leetcode.medium;
 
 import mx.grekz.leetcode.helpers.ListNode;
